rewritted InfoBar from js to ts
