+++ src/components/InfoBar/index.ts
